# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "B2C Marketing Automation API",
    "description": "Complete REST API for the B2C Marketing Automation Platform. Manage contacts, workflows, deliveries, segments, audiences, custom objects, and more.",
    "version": "2.0.0",
    "contact": {
      "name": "Platform Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/orchestration/{campaignId}": {
      "get": {
        "tags": ["Orchestration"],
        "summary": "Get orchestration",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Orchestration data" } }
      },
      "post": {
        "tags": ["Orchestration"],
        "summary": "Save orchestration",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Orchestration saved" } }
      }
    },
    "/orchestration/{campaignId}/execute": {
      "post": {
        "tags": ["Orchestration"],
        "summary": "Execute orchestration",
        "description": "Run the campaign with the orchestration flow.",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Execution started" } }
      }
    },
    "/orchestration/{campaignId}/waits": {
      "get": {
        "tags": ["Orchestration"],
        "summary": "Monitor waiting signals",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Waiting signal list" } }
      }
    },
    "/orchestration/{campaignId}/signal": {
      "post": {
        "tags": ["Orchestration"],
        "summary": "Send external signal",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Signal received" } }
      }
    },
    "/orchestration/{campaignId}/preview/schema": {
      "get": {
        "tags": ["Orchestration"],
        "summary": "Preview schema",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Schema preview" } }
      }
    },
    "/orchestration/{campaignId}/preview/results": {
      "get": {
        "tags": ["Orchestration"],
        "summary": "Preview results",
        "parameters": [{ "name": "campaignId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Results preview" } }
      }
    },
...
    "/segments": {
      "get": {
        "tags": ["Segments"],
        "summary": "List segments",
        "responses": { "200": { "description": "List of segments" } }
      },
      "post": {
        "tags": ["Segments"],
        "summary": "Create segment",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SegmentInput" } } } },
        "responses": { "201": { "description": "Segment created" } }
      }
    },
...
    "/segments/{id}": {
      "get": {
        "tags": ["Segments"],
        "summary": "Get segment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Segment details" } }
      },
      "put": {
        "tags": ["Segments"],
        "summary": "Update segment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Segment updated" } }
      },
      "delete": {
        "tags": ["Segments"],
        "summary": "Delete segment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Segment deleted" } }
      }
    },
...
    "/audiences/{id}": {
      "get": { "tags": ["Audiences"], "summary": "Get audience", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Audience details" } } },
      "put": { "tags": ["Audiences"], "summary": "Update audience", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["Audiences"], "summary": "Delete audience", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Deleted" } } }
    },
...
  },
  "components": {
    "schemas": {
...
      "SegmentInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "segment_type": { "type": "string" },
          "rules": { "type": "array", "items": { "type": "object" } }
        },
        "required": ["name"]
      },
...
    }
  }
}